// src/types/educationType.ts
import { ApplicationStatus } from "@/constants/base-constant";
import { DegreeName } from "@/constants/education-constan";

export type Education = {
    id:string;
    graduatedAt: string;
    status: ApplicationStatus;
    trainingName: string;
    majorName: string;
    degreeValue: number;
    degreeName: DegreeName;
    lastModify:Date
};